refactor: split wallet component into different components
